piprequirements.txt
python-telegram-bot
requests
os
loging
csv
arparse
json
time
datetime
threading
requests
gtts
python-vlc

